refactor: make PaymentAware read only
