+++ Tests/AdyenTests/Adyen Tests/Components/Gift Card/GiftCardComponentTests.swift
@@ -28,6 +28,8 @@ class GiftCardComponentTests: XCTestCase {
 
     var paymentMethod: GiftCardPaymentMethod!
 
+    let amountToPay = Amount(value: 100, currencyCode: "EUR")
+
     var errorView: FormErrorItemView? {
         sut.viewController.view.findView(with: "AdyenCard.GiftCardComponent.errorItem")
     }
@@ -54,6 +56,7 @@ class GiftCardComponentTests: XCTestCase {
 
         sut = GiftCardComponent(paymentMethod: paymentMethod,
                                 context: context,
+                                amount: amountToPay,
                                 publicKeyProvider: publicKeyProvider)
         delegateMock = PaymentComponentDelegateMock()
         sut.delegate = delegateMock
@@ -125,8 +128,6 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 10, currencyCode: "EUR"), countryCode: "NL")
-
         delegateMock.onDidFail = { error, component in
             XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
         }
@@ -165,14 +166,12 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 10, currencyCode: "USD"), countryCode: "US")
-
         delegateMock.onDidFail = { error, component in
             XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
         }
 
         let onCheckBalanceExpectation = expectation(description: "Expect partialPaymentDelegate.onCheckBalance to be called.")
-        let balance = Balance(availableAmount: .init(value: 200, currencyCode: "EUR"), transactionLimit: .init(value: 1000, currencyCode: "EUR"))
+        let balance = Balance(availableAmount: .init(value: 200, currencyCode: "USD"), transactionLimit: .init(value: 1000, currencyCode: "USD"))
         partialPaymentDelegate.onCheckBalance = { _, completion in
             completion(.success(balance))
             onCheckBalanceExpectation.fulfill()
@@ -205,8 +204,6 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 10, currencyCode: "EUR"), countryCode: "US")
-
         delegateMock.onDidFail = { error, component in
             XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
         }
@@ -236,44 +233,6 @@ class GiftCardComponentTests: XCTestCase {
         waitForExpectations(timeout: 10, handler: nil)
     }
 
-    func testMissingPaymentObject() throws {
-
-        let publicKeyProviderExpectation = expectation(description: "Expect publicKeyProvider to be called.")
-        publicKeyProviderExpectation.expectedFulfillmentCount = 2
-        publicKeyProvider.onFetch = { completion in
-            publicKeyProviderExpectation.fulfill()
-            completion(.success(Dummy.publicKey))
-        }
-
-        delegateMock.onDidFail = { error, component in
-            XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
-        }
-
-        let onCheckBalanceExpectation = expectation(description: "Expect partialPaymentDelegate.onCheckBalance to be called.")
-        let balance = Balance(availableAmount: .init(value: 200, currencyCode: "EUR"), transactionLimit: .init(value: 1000, currencyCode: "EUR"))
-        partialPaymentDelegate.onCheckBalance = { _, completion in
-            completion(.success(balance))
-            onCheckBalanceExpectation.fulfill()
-        }
-
-        UIApplication.shared.keyWindow?.rootViewController = sut.viewController
-
-        wait(for: .milliseconds(300))
-
-        XCTAssertTrue(errorView!.isHidden)
-
-        populate(cardNumber: "60643650100000000000", pin: "73737")
-
-        payButtonItemViewButton?.sendActions(for: .touchUpInside)
-
-        wait(for: .seconds(1))
-
-        XCTAssertFalse(errorView!.isHidden)
-        XCTAssertEqual(sut.errorItem.message, "An unknown error occurred")
-
-        waitForExpectations(timeout: 10, handler: nil)
-    }
-
     func testEnoughBalanceIsAvailableWithNilReadyToSubmitDelegate() throws {
 
         let publicKeyProviderExpectation = expectation(description: "Expect publicKeyProvider to be called.")
@@ -283,7 +242,6 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 100, currencyCode: "EUR"), countryCode: "NL")
         sut.readyToSubmitComponentDelegate = nil
 
         delegateMock.onDidFail = { error, component in
@@ -303,7 +261,6 @@ class GiftCardComponentTests: XCTestCase {
 
         let onSubmitExpectation = expectation(description: "Expect delegateMock.onDidSubmit to be called.")
         delegateMock.onDidSubmit = { data, component in
-            XCTAssertEqual(data.amount, self.sut.payment!.amount)
             XCTAssertNil(data.order)
             XCTAssertTrue(component === self.sut)
             onSubmitExpectation.fulfill()
@@ -335,8 +292,6 @@ class GiftCardComponentTests: XCTestCase {
             publicKeyProviderExpectation.fulfill()
             completion(.success(Dummy.publicKey))
         }
-
-        sut.payment = Payment(amount: .init(value: 100, currencyCode: "EUR"), countryCode: "NL")
 
         delegateMock.onDidFail = { error, component in
             XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
@@ -390,8 +345,6 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 100, currencyCode: "EUR"), countryCode: "NL")
-
         delegateMock.onDidFail = { error, component in
             XCTFail("delegateMock.onDidFail shouldn't be reported back to merchant.")
         }
@@ -413,7 +366,6 @@ class GiftCardComponentTests: XCTestCase {
         let onSubmitExpectation = expectation(description: "Expect delegateMock.onDidSubmit to be called.")
         delegateMock.onDidSubmit = { data, component in
             XCTAssertEqual(data.order, expectedOrder)
-            XCTAssertEqual(data.amount, .init(value: 50, currencyCode: "EUR"))
             onSubmitExpectation.fulfill()
         }
 
@@ -448,7 +400,6 @@ class GiftCardComponentTests: XCTestCase {
             completion(.success(Dummy.publicKey))
         }
 
-        sut.payment = Payment(amount: .init(value: 100, currencyCode: "EUR"), countryCode: "NL")
         sut.order = PartialPaymentOrder(pspReference: "pspreference", orderData: "data")
 
         delegateMock.onDidFail = { error, component in
@@ -470,7 +421,6 @@ class GiftCardComponentTests: XCTestCase {
         let onSubmitExpectation = expectation(description: "Expect delegateMock.onDidSubmit to be called.")
         delegateMock.onDidSubmit = { data, component in
             XCTAssertEqual(data.order, expectedOrder)
-            XCTAssertEqual(data.amount, .init(value: 50, currencyCode: "EUR"))
             onSubmitExpectation.fulfill()
         }
 
@@ -504,8 +454,6 @@ class GiftCardComponentTests: XCTestCase {
             publicKeyProviderExpectation.fulfill()
             completion(.success(Dummy.publicKey))
         }
-
-        sut.payment = Payment(amount: .init(value: 100, currencyCode: "EUR"), countryCode: "NL")
 
         let onDidFailExpectation = expectation(description: "Expect delegateMock.onDidFail to be called.")
         delegateMock.onDidFail = { error, component in
